fix: rename offer-template
