+++ src/components/trip/events/offer-template.js
